+++ src/main/java/cs4224/entities/OrderLine.java
@@ -17,7 +17,7 @@ import java.math.BigDecimal;
 public class OrderLine {
 
     @JsonProperty("ol_i_id")
-    private Integer itemId;
+    private long itemId;
 
     @JsonProperty("ol_quantity")
     private BigDecimal quantity;
